Apply Factory Design Pattern to Player, Card and Deck Creation
Body:
This commit applies the Factory Design Pattern to the creation of Player, Card and Deck objects in the game.

- Created PlayerFactory, CardFactory and DeckFactory classes to encapsulate the creation logic of Player, Card and Deck objects respectively.

- Refactored PlayerManager class to use PlayerFactory for player creation.

- Refactored Player class to use DeckFactory for deck creation.

- Refactored CardLibrary class to use CardFactory for card creation

These changes make the code more flexible and easier to maintain, as adding new types of players, cards or decks in the future will only require modifications in the respective factory classes.

If the game were to support different types of game modes or levels, a Factory could be used to create these as well.